Reduce redundancy in conductor calculations
