+++ examples/transport_examples/example02/main_conductor.py
@@ -65,10 +65,8 @@ def main() -> None:
     # Compute and write transport observables by physics stage.
     transport.compute_leads_self_energy(write=True)
     transport.compute_greens_functions(write=True)
-    transmission = transport.compute_transmission(write=True)
-    dos = transport.compute_dos(write=True)
-    print('Transmission shape:', transmission.shape)
-    print('DOS shape:', dos.shape)
+    transport.compute_transmission()
+    transport.compute_dos()
 
 
 if __name__ == '__main__':
